Fix roomCode bug
However, for some reason when a user is given a link with a roomCode, the url params aren't read on the first load.  Not sure if this is a issue in dev mode but useLocation().search is an empty string on first load but refreshing the page makes it work.